api/requirements.txt
fastapi
uvicorn[standard]
chromadb
pandas
pyarrow
sentence-transformers
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
pydantic
httpx
python-dotenv